+++ __mocks__/axios.js
@@ -1,0 +1,7 @@
+const axios = {
+    get: jest.fn(),
+    post: jest.fn(),
+    create: jest.fn(() => axios)
+};
+
+module.exports = axios;